# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Resume {
  id: number;
  /** @nullable */
  userId?: string | null;
  rawText: string;
  /** @nullable */
  fileUrl?: string | null;
  iteration: number;
  pipelineStage: string;
  /** @nullable */
  matchScore?: number | null;
  createdAt: string;
  /** @nullable */
  jobDescription?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  company?: string | null;
}

export interface StageStatus {
  name: string;
  status: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface PipelineStatus {
  resumeId: number;
  stage: string;
  stages: StageStatus[];
  /** @nullable */
  matchScore?: number | null;
}

export interface ExtractedSkill {
  id: number;
  skillName: string;
  category: string;
  confidenceScore: number;
}

export interface ExtractionResult {
  resumeId: number;
  skills: ExtractedSkill[];
  /** @nullable */
  experienceYears?: number | null;
  jobTitles: string[];
  education: string[];
  tools: string[];
}

export interface SkillGap {
  id: number;
  missingSkill: string;
  importanceLevel: string;
  /** @nullable */
  suggestion?: string | null;
}

export interface Suggestion {
  id: number;
  gapId: number;
  missingSkill: string;
  importanceLevel: string;
  bullets: string[];
  applied: boolean;
}

export interface RewriteResult {
  resumeId: number;
  originalText: string;
  rewrittenText: string;
  versionNumber: number;
}

export interface DiffChunk {
  type: string;
  text: string;
}

export interface DiffResult {
  resumeId: number;
  originalText: string;
  rewrittenText: string;
  changeCount: number;
  diffs: DiffChunk[];
}

export interface ChecklistItem {
  item: string;
  passed: boolean;
  /** @nullable */
  detail?: string | null;
}

export interface ValidationResult {
  resumeId: number;
  atsScore: number;
  keywordScore: number;
  formattingScore: number;
  skillMatchScore: number;
  checklist: ChecklistItem[];
}

export interface ResumeVersion {
  id: number;
  resumeId: number;
  versionNumber: number;
  content: string;
  /** @nullable */
  score?: number | null;
  createdAt: string;
}

export interface MatchScore {
  resumeId: number;
  score: number;
  label: string;
}

export type UploadResumeBody = {
  rawText: string;
  jobDescription: string;
  jobTitle?: string;
  company?: string;
  portfolioUrl?: string;
};