# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Resume Analyzer API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: resumes
    description: Resume operations
  - name: pipeline
    description: AI pipeline operations
  - name: versions
    description: Resume version history
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /resume/upload:
    post:
      operationId: uploadResume
      tags: [resumes]
      summary: Upload a resume with job description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rawText:
                  type: string
                jobDescription:
                  type: string
                jobTitle:
                  type: string
                company:
                  type: string
                portfolioUrl:
                  type: string
              required:
                - rawText
                - jobDescription
      responses:
        "201":
          description: Resume uploaded and pipeline started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "400":
          description: Invalid input

  /resume:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List all resumes
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"

  /resume/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Resume not found

  /resume/{id}/status:
    get:
      operationId: getResumeStatus
      tags: [pipeline]
      summary: Get current pipeline status for a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pipeline status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStatus"
        "404":
          description: Resume not found

  /resume/{id}/extraction:
    get:
      operationId: getExtraction
      tags: [pipeline]
      summary: Get extracted skills and experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extracted data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionResult"
        "404":
          description: Not found

  /resume/{id}/gaps:
    get:
      operationId: getGaps
      tags: [pipeline]
      summary: Get skill gaps
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skill gaps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillGap"
        "404":
          description: Not found

  /resume/{id}/suggestions:
    get:
      operationId: getSuggestions
      tags: [pipeline]
      summary: Get improvement suggestions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggestions per gap
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Suggestion"
        "404":
          description: Not found

  /resume/{id}/rewrite:
    get:
      operationId: getRewrite
      tags: [pipeline]
      summary: Get the rewritten resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rewritten resume content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewriteResult"
        "404":
          description: Not found

  /resume/{id}/diff:
    get:
      operationId: getDiff
      tags: [pipeline]
      summary: Get before/after diff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Diff data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiffResult"
        "404":
          description: Not found

  /resume/{id}/validation:
    get:
      operationId: getValidation
      tags: [pipeline]
      summary: Get ATS validation results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"
        "404":
          description: Not found

  /resume/{id}/iterate:
    post:
      operationId: iterateResume
      tags: [pipeline]
      summary: Trigger a new pipeline run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New pipeline run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found

  /resume/{id}/versions:
    get:
      operationId: getVersions
      tags: [versions]
      summary: Get all versions of a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResumeVersion"
        "404":
          description: Not found

  /resume/{id}/match-score:
    get:
      operationId: getMatchScore
      tags: [pipeline]
      summary: Get overall resume-to-job match score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchScore"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Resume:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        rawText:
          type: string
        fileUrl:
          type: ["string", "null"]
        iteration:
          type: integer
        pipelineStage:
          type: string
        matchScore:
          type: ["number", "null"]
        createdAt:
          type: string
        jobDescription:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
      required:
        - id
        - rawText
        - iteration
        - pipelineStage
        - createdAt

    PipelineStatus:
      type: object
      properties:
        resumeId:
          type: integer
        stage:
          type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/StageStatus"
        matchScore:
          type: ["number", "null"]
      required:
        - resumeId
        - stage
        - stages

    StageStatus:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - name
        - status

    ExtractionResult:
      type: object
      properties:
        resumeId:
          type: integer
        skills:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedSkill"
        experienceYears:
          type: ["number", "null"]
        jobTitles:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
      required:
        - resumeId
        - skills
        - jobTitles
        - education
        - tools

    ExtractedSkill:
      type: object
      properties:
        id:
          type: integer
        skillName:
          type: string
        category:
          type: string
        confidenceScore:
          type: number
      required:
        - id
        - skillName
        - category
        - confidenceScore

    SkillGap:
      type: object
      properties:
        id:
          type: integer
        missingSkill:
          type: string
        importanceLevel:
          type: string
        suggestion:
          type: ["string", "null"]
      required:
        - id
        - missingSkill
        - importanceLevel

    Suggestion:
      type: object
      properties:
        id:
          type: integer
        gapId:
          type: integer
        missingSkill:
          type: string
        importanceLevel:
          type: string
        bullets:
          type: array
          items:
            type: string
        applied:
          type: boolean
      required:
        - id
        - gapId
        - missingSkill
        - importanceLevel
        - bullets
        - applied

    RewriteResult:
      type: object
      properties:
        resumeId:
          type: integer
        originalText:
          type: string
        rewrittenText:
          type: string
        versionNumber:
          type: integer
      required:
        - resumeId
        - originalText
        - rewrittenText
        - versionNumber

    DiffResult:
      type: object
      properties:
        resumeId:
          type: integer
        originalText:
          type: string
        rewrittenText:
          type: string
        changeCount:
          type: integer
        diffs:
          type: array
          items:
            $ref: "#/components/schemas/DiffChunk"
      required:
        - resumeId
        - originalText
        - rewrittenText
        - changeCount
        - diffs

    DiffChunk:
      type: object
      properties:
        type:
          type: string
        text:
          type: string
      required:
        - type
        - text

    ValidationResult:
      type: object
      properties:
        resumeId:
          type: integer
        atsScore:
          type: number
        keywordScore:
          type: number
        formattingScore:
          type: number
        skillMatchScore:
          type: number
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
      required:
        - resumeId
        - atsScore
        - keywordScore
        - formattingScore
        - skillMatchScore
        - checklist

    ChecklistItem:
      type: object
      properties:
        item:
          type: string
        passed:
          type: boolean
        detail:
          type: ["string", "null"]
      required:
        - item
        - passed

    ResumeVersion:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        versionNumber:
          type: integer
        content:
          type: string
        score:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - resumeId
        - versionNumber
        - content
        - createdAt

    MatchScore:
      type: object
      properties:
        resumeId:
          type: integer
        score:
          type: number
        label:
          type: string
      required:
        - resumeId
        - score
        - label